% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{path_downloads}
\alias{path_downloads}
\title{Help to create a path to downloads}
\usage{
path_downloads(file = NULL)
}
\arguments{
\item{file}{File name.}
}
\value{
A path.
}
\description{
Help to create a path to downloads
}
\examples{
\dontrun{
path_downloads()
path_downloads("fgeo_1.0.0.tgz")
}
}
