% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path_src}
\alias{path_src}
\alias{path_win}
\alias{path_mac}
\title{Create paths to packages in a CRAN-like repository.}
\usage{
path_src()

path_win()

path_mac(mac_version = "el-capitan")
}
\arguments{
\item{mac_version}{Versoin of mac the package was build on.}
}
\value{
Path
}
\description{
Create paths to packages in a CRAN-like repository.
}
\examples{
\dontrun{
path_src()
path_win()
path_mac()
}
}
